requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
google-generativeai==0.3.1
python-dotenv==1.0.0
